PR-005: move EmbeddingRepository trait into api module
