+++ src/api/mod.rs
@@ -1,2 +1,4 @@
 pub mod embedding;
 pub mod types;
+
+pub use embedding::EmbeddingRepository;
