+++ src/repositories.rs
@@ -1,8 +1,6 @@
 // Module declarations
-mod embedding_repository;
 mod memory_repository;
 mod vector_memory_repository;
 
-pub use embedding_repository::EmbeddingRepository;
 pub(crate) use memory_repository::MemoryRepository;
 pub(crate) use vector_memory_repository::VectorMemoryRepository;
